chore: move reusable functions from news.mjs to news-helpers.mjs, linting fixes
