+++ scripts/news-helpers.d.ts
